+++ src/app/helpers/TAKE_GENERATOR_TYPES.js
@@ -1,0 +1,9 @@
+const TAKE_GENERATOR_TYPES = {
+  TAKE_EVERY: '@@sagaTester__takeEvery__',
+  TAKE_LATEST: '@@sagaTester__takeLatest__',
+  TAKE_LEADING: '@@sagaTester__takeLeading__',
+  THROTTLE: '@@sagaTester__throttle__',
+  DEBOUNCE: '@@sagaTester__debounce__',
+};
+
+export default TAKE_GENERATOR_TYPES;